lenovo/yoga/7/pro-x: add 14IAH7; create shared for 14ARH7 and 14IAH7
`nixos-hardware` already had Lenovo 14ARH7 (Lenovo Yoga Slim 7 Pro X) model
profiles. This an AMD device with an NVIDIA dedicated GPU. There is a profile
for integrated graphics only and another for hybrid graphics.

There is an Intel version of the same device, the Lenovo 14IAH7
(Lenovo Yoga Slim 7i Pro X). Seeing as they are primarily the same device, save
the difference in processor under the Lenovo Yoga Slim 7 Pro X "line", the
listed was done to share values between the AMD version and the Intel version
that was added in this commit. It follows the same pattern of having a profile
for integrated graphics only and another for hybrid graphics.

The values of the 14ARH7 modules were not changed, just moved around.
The 14ARH7 module names were left as is to not cause unnecessary
breakages.

- Create 'pro-x' subsection in `lenovo/yoga/7`
- Pull content from '14ARH7' profiles to share with '14IAH7' in `lenovo/yoga/7/pro-x/shared`
- Strip '14ARH7' profiles to use shared modules and only have unique values
- Create profiles for '14IAH7'
- Update `flake.nix` and `README.md`